crate refactor consistency
