+++ fighters/richter/src/acmd/mod.rs
@@ -1,18 +1,19 @@
 use super::*;
-mod aerials;
+
+mod ground;
 mod tilts;
-mod other;
 mod smashes;
+mod aerials;
 mod specials;
 mod throws;
-mod ground;
+mod other;
 
 pub fn install(agent: &mut Agent) {
-    aerials::install(agent);
+    ground::install(agent);
     tilts::install(agent);
-    other::install(agent);
     smashes::install(agent);
+    aerials::install(agent);
     specials::install(agent);
     throws::install(agent);
-    ground::install(agent);
+    other::install(agent);
 }
