+++ fighters/richter/src/acmd/smashes.rs
@@ -1,34 +1,6 @@
 use super::*;
 
-unsafe extern "C" fn game_attacks4hi(agent: &mut L2CAgentBase) {
-    let lua_state = agent.lua_state_agent;
-    let boma = agent.boma();
-    frame(lua_state, 6.0);
-    if is_excute(agent) {
-        WorkModule::on_flag(boma, *FIGHTER_STATUS_ATTACK_FLAG_START_SMASH_HOLD);
-    }
-    frame(lua_state, 24.0);
-    for _ in 0..5 {
-        if is_excute(agent) {
-            ATTACK(agent, 0, 0, Hash40::new("top"), 1.0, 365, 100, 30, 0, 3.8, 0.0, 13.5, 44.5, Some(0.0), Some(7.6), Some(7.7), 0.75, 1.0, *ATTACK_SETOFF_KIND_OFF, *ATTACK_LR_CHECK_F, false, 1.5, 0.0, 0, false, false, false, false, true, *COLLISION_SITUATION_MASK_GA, *COLLISION_CATEGORY_MASK_ALL, *COLLISION_PART_MASK_ALL, false, Hash40::new("collision_attr_fire"), *ATTACK_SOUND_LEVEL_L, *COLLISION_SOUND_ATTR_RICHTER_WHIP, *ATTACK_REGION_WHIP);
-        }
-        wait(lua_state, 1.0);
-        if is_excute(agent) {
-            AttackModule::clear_all(boma);
-        }
-        wait(lua_state, 1.0);
-    }
-    if is_excute(agent) {
-        ATTACK(agent, 1, 0, Hash40::new("top"), 10.0, 361, 85, 0, 60, 4.0, 0.0, 13.8, 39.0, Some(0.0), Some(7.6), Some(7.5), 0.75, 1.0, *ATTACK_SETOFF_KIND_OFF, *ATTACK_LR_CHECK_F, false, 2.0, 0.0, 0, false, false, false, false, true, *COLLISION_SITUATION_MASK_GA, *COLLISION_CATEGORY_MASK_ALL, *COLLISION_PART_MASK_ALL, false, Hash40::new("collision_attr_fire"), *ATTACK_SOUND_LEVEL_L, *COLLISION_SOUND_ATTR_RICHTER_WHIP, *ATTACK_REGION_WHIP);
-        ATTACK(agent, 0, 0, Hash40::new("top"), 15.0, 50, 85, 0, 60, 4.0, 0.0, 13.7, 45.0, Some(0.0), Some(13.5), Some(43.0), 1.0, 1.0, *ATTACK_SETOFF_KIND_THRU, *ATTACK_LR_CHECK_F, false, 3.0, 0.0, 0, false, false, false, false, true, *COLLISION_SITUATION_MASK_GA, *COLLISION_CATEGORY_MASK_ALL, *COLLISION_PART_MASK_ALL, false, Hash40::new("collision_attr_fire"), *ATTACK_SOUND_LEVEL_L, *COLLISION_SOUND_ATTR_FIRE, *ATTACK_REGION_WHIP);
-    }
-    wait(lua_state, 2.0);
-    if is_excute(agent) {
-        AttackModule::clear_all(boma);
-    }
-}
-
-unsafe extern "C" fn game_sttacks4(agent: &mut L2CAgentBase) {
+unsafe extern "C" fn game_attacks4(agent: &mut L2CAgentBase) {
     let lua_state = agent.lua_state_agent;
     let boma = agent.boma();
     frame(lua_state, 6.0);
@@ -157,6 +129,34 @@ unsafe extern "C" fn effect_attacks4(agent: &mut L2CAgentBase) {
     frame(lua_state, 35.0);
     if is_excute(agent) {
         EFFECT_OFF_KIND(agent, Hash40::new("sys_firebar_trace"), false, true);
+    }
+}
+
+unsafe extern "C" fn game_attacks4hi(agent: &mut L2CAgentBase) {
+    let lua_state = agent.lua_state_agent;
+    let boma = agent.boma();
+    frame(lua_state, 6.0);
+    if is_excute(agent) {
+        WorkModule::on_flag(boma, *FIGHTER_STATUS_ATTACK_FLAG_START_SMASH_HOLD);
+    }
+    frame(lua_state, 24.0);
+    for _ in 0..5 {
+        if is_excute(agent) {
+            ATTACK(agent, 0, 0, Hash40::new("top"), 1.0, 365, 100, 30, 0, 3.8, 0.0, 13.5, 44.5, Some(0.0), Some(7.6), Some(7.7), 0.75, 1.0, *ATTACK_SETOFF_KIND_OFF, *ATTACK_LR_CHECK_F, false, 1.5, 0.0, 0, false, false, false, false, true, *COLLISION_SITUATION_MASK_GA, *COLLISION_CATEGORY_MASK_ALL, *COLLISION_PART_MASK_ALL, false, Hash40::new("collision_attr_fire"), *ATTACK_SOUND_LEVEL_L, *COLLISION_SOUND_ATTR_RICHTER_WHIP, *ATTACK_REGION_WHIP);
+        }
+        wait(lua_state, 1.0);
+        if is_excute(agent) {
+            AttackModule::clear_all(boma);
+        }
+        wait(lua_state, 1.0);
+    }
+    if is_excute(agent) {
+        ATTACK(agent, 1, 0, Hash40::new("top"), 10.0, 361, 85, 0, 60, 4.0, 0.0, 13.8, 39.0, Some(0.0), Some(7.6), Some(7.5), 0.75, 1.0, *ATTACK_SETOFF_KIND_OFF, *ATTACK_LR_CHECK_F, false, 2.0, 0.0, 0, false, false, false, false, true, *COLLISION_SITUATION_MASK_GA, *COLLISION_CATEGORY_MASK_ALL, *COLLISION_PART_MASK_ALL, false, Hash40::new("collision_attr_fire"), *ATTACK_SOUND_LEVEL_L, *COLLISION_SOUND_ATTR_RICHTER_WHIP, *ATTACK_REGION_WHIP);
+        ATTACK(agent, 0, 0, Hash40::new("top"), 15.0, 50, 85, 0, 60, 4.0, 0.0, 13.7, 45.0, Some(0.0), Some(13.5), Some(43.0), 1.0, 1.0, *ATTACK_SETOFF_KIND_THRU, *ATTACK_LR_CHECK_F, false, 3.0, 0.0, 0, false, false, false, false, true, *COLLISION_SITUATION_MASK_GA, *COLLISION_CATEGORY_MASK_ALL, *COLLISION_PART_MASK_ALL, false, Hash40::new("collision_attr_fire"), *ATTACK_SOUND_LEVEL_L, *COLLISION_SOUND_ATTR_FIRE, *ATTACK_REGION_WHIP);
+    }
+    wait(lua_state, 2.0);
+    if is_excute(agent) {
+        AttackModule::clear_all(boma);
     }
 }
 
@@ -326,13 +326,13 @@ unsafe extern "C" fn game_attacklw4(agent: &mut L2CAgentBase) {
 }
 
 pub fn install(agent: &mut Agent) {
+    agent.acmd("game_attacks4", game_attacks4);
+    agent.acmd("effect_attacks4", effect_attacks4);
+    agent.acmd("sound_attacks4", sound_attacks4);
+
     agent.acmd("game_attacks4hi", game_attacks4hi);
     agent.acmd("effect_attacks4hi", effect_attacks4);
     agent.acmd("sound_attacks4hi", sound_attacks4);
-
-    agent.acmd("game_attacks4", game_sttacks4);
-    agent.acmd("effect_attacks4", effect_attacks4);
-    agent.acmd("sound_attacks4", sound_attacks4);
 
     agent.acmd("game_attacks4lw", game_attacks4lw);
     agent.acmd("effect_attacks4lw", effect_attacks4);
