+++ fighters/richter/src/acmd/specials.rs
@@ -170,7 +170,7 @@ unsafe extern "C" fn sound_specials1(agent: &mut L2CAgentBase) {
 unsafe extern "C" fn expression_specials1(agent: &mut L2CAgentBase) {
     let lua_state = agent.lua_state_agent;
     let boma = agent.boma();
-    if macros::is_excute(agent) {
+    if is_excute(agent) {
         slope!(agent, *MA_MSC_CMD_SLOPE_SLOPE, *SLOPE_STATUS_LR);
     }
     frame(lua_state, 12.0);
