+++ fighters/richter/src/acmd/tilts.rs
@@ -1,6 +1,6 @@
 use super::*;
 
-unsafe extern "C" fn game__attacks3(agent: &mut L2CAgentBase) {
+unsafe extern "C" fn game_attacks3(agent: &mut L2CAgentBase) {
     let lua_state = agent.lua_state_agent;
     let boma = agent.boma();
     frame(lua_state, 12.0);
@@ -210,16 +210,16 @@ unsafe extern "C" fn effect_attackhi3(agent: &mut L2CAgentBase) {
 unsafe extern "C" fn expression_attackhi3(agent: &mut L2CAgentBase) {
     let lua_state = agent.lua_state_agent;
     let boma = agent.boma();
-    if macros::is_excute(agent) {
+    if is_excute(agent) {
         slope!(agent, *MA_MSC_CMD_SLOPE_SLOPE, *SLOPE_STATUS_LR);
     }
     frame(lua_state, 8.0);
-    if macros::is_excute(agent) {
+    if is_excute(agent) {
         ControlModule::set_rumble(agent.module_accessor, Hash40::new("rbkind_nohitm"), 0, false, *BATTLE_OBJECT_ID_INVALID as u32);
     }
     frame(lua_state, 10.0);
-    if macros::is_excute(agent) {
-        macros::RUMBLE_HIT(agent, Hash40::new("rbkind_attackm"), 0);
+    if is_excute(agent) {
+        RUMBLE_HIT(agent, Hash40::new("rbkind_attackm"), 0);
     }
 }
 
@@ -308,7 +308,7 @@ unsafe extern "C" fn game_attacklw32(agent: &mut L2CAgentBase) {
 }
 
 pub fn install(agent: &mut Agent) {
-    agent.acmd("game_attacks3", game__attacks3);
+    agent.acmd("game_attacks3", game_attacks3);
     agent.acmd("effect_attacks3", effect_attacks3);
     agent.acmd("sound_attacks3", sound_attacks3);
     agent.acmd("expression_attacks3", expression_attacks3);
