+++ fighters/richter/src/lib.rs
@@ -4,10 +4,11 @@
 
 pub mod acmd;
 
-pub mod status;
 pub mod opff;
+pub mod status;
 
 // articles
+
 mod axe;
 mod whip;
 
@@ -41,12 +42,13 @@ use utils::{
     consts::*,
 };
 use smashline::*;
+#[macro_use] extern crate smash_script;
 
 pub fn install() {
     let agent = &mut Agent::new("richter");
     acmd::install(agent);
-    status::install(agent);
     opff::install(agent);
+    status::install(agent);
     agent.install();
 
     axe::install();
