+++ fighters/richter/src/status/attacks3.rs
@@ -1,5 +1,4 @@
 use super::*;
-use globals::*;
 
 // FIGHTER_STATUS_KIND_ATTACK_S3
