+++ fighters/richter/src/status/mod.rs
@@ -1,4 +1,5 @@
 use super::*;
+use globals::*;
 
 mod attacklw3;
 mod attacks3;
